added:
  daemon/system/System/ignore.hpp
  daemon/system/System/ignore.t.cpp
  doc/devel/creating_filters.txt
  doc/devel/creating_inputs.txt
  doc/devel/creating_triggers.txt
  doc/example_configs/
  doc/example_configs/example1.xml
  doc/example_configs/minimal.xml
modified:
  daemon/commons/Commons/Filesystem/openElement.cpp
  daemon/commons/Commons/Filesystem/openElement.t.cpp
  daemon/commons/Commons/SharedPtrNotNULL.hpp
  daemon/core/Core/Processors.cpp
  daemon/core/Core/Sources.cpp
  daemon/filter/Filter/Simple/Strategy.hpp
  daemon/input/Input/Thread.cpp
  daemon/inputfile/Input/File/ParserCallback.cpp
  daemon/inputfile/Input/File/ParserCallback.t.cpp
  daemon/rfcio/RFCIO/FromXML.cpp
  daemon/system/System/ScopedPtrCustom.hpp
  daemon/system/System/SharedPtrNotNULL.hpp
  daemon/system/System/SharedPtrNotNULL.t.cpp
  daemon/trigger/Trigger/Compose/Full.t.cpp
  daemon/trigger/Trigger/Compose/LinkWUI.cpp
  daemon/trigger/Trigger/Compose/LinkWUI.t.cpp
  daemon/trigger/Trigger/Compose/Summary.t.cpp
  data_model/postgres/8.3/06_drop_meta_alerts.sql
  data_model/postgres/8.3/11_create_wui.sql
  data_model/postgres/8.3/12_create_alerts.sql
  data_model/postgres/8.3/13_create_meta_alerts.sql
  data_model/postgres/8.3/14_create_processor_states.sql
  data_model/postgres/8.4/06_drop_meta_alerts.sql
  data_model/postgres/8.4/11_create_wui.sql
  data_model/postgres/8.4/12_create_alerts.sql
  data_model/postgres/8.4/13_create_meta_alerts.sql
  data_model/postgres/8.4/14_create_processor_states.sql
  wui/site/protected/common/DataTableTemplate.tpl
pending merges:
  Bartek 'BaSz' Szu... 2010-11-17 commente aded.
    Bartek 'BaSz' Szu... 2010-11-17 updated version of sharedptrnn added.
    Bartek 'BaSz' Szu... 2010-11-17 changed some important logs from DEBUG to INFO so that they'll be present in release mode as well;
    Bartek 'BaSz' Szu... 2010-11-16 updated DB schemas for PSQL to use 64-bit integers as indexes.
    Bartek 'BaSz' Szu... 2010-11-16 example config files added.
    Bartek 'BaSz' Szu... 2010-11-16 info on changing ID is now WARN not ERROR.
    Bartek 'BaSz' Szu... 2010-11-16 fixed issue with reading files that have no write-access.
    Bartek 'BaSz' Szu... 2010-11-16 added missing test on openElement call;
    Bartek 'BaSz' Szu... 2010-11-16 typo fixed.
    Bartek 'BaSz' Szu... 2010-11-16 added more TODOs on documentation.
    Bartek 'BaSz' Szu... 2010-11-15 fixed tests for changed URLs.
    Bartek 'BaSz' Szu... 2010-11-15 new, development version of System library included.
    Bartek 'BaSz' Szu... 2010-11-15 page updated to new WUI.
    Bartek 'BaSz' Szu... 2010-11-09 TODO added.
    Bartek 'BaSz' Szu... 2010-11-09 [merge] rebase.

+++ data_model/postgres/8.3/06_drop_meta_alerts.sql
@@ -9,13 +9,13 @@ DROP TABLE    meta_alerts_tree;
 
 DROP FUNCTION ensure_leafs_are_unique();
 
-DROP FUNCTION link_meta_alerts(int, int);
+DROP FUNCTION link_meta_alerts(bigint, bigint);
 
-DROP FUNCTION get_parents(int);
-DROP FUNCTION get_parents_impl(int);
+DROP FUNCTION get_parents(bigint);
+DROP FUNCTION get_parents_impl(bigint);
 
-DROP FUNCTION get_children(int);
-DROP FUNCTION get_children_impl(int);
+DROP FUNCTION get_children(bigint);
+DROP FUNCTION get_children_impl(bigint);
 
 DROP TABLE    meta_alerts;
 DROP SEQUENCE meta_alerts_id_seq;
